# docs/contracts/staff-bff-documents.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Aurora Staff BFF — Documents/OCR contract",
    "version": "v1",
    "description": "Deterministic FE contract fixture for the source-backed Documents/OCR upload and intake workflow. It is intentionally sanitized: signed URL values, tenant IDs, storage credentials and authentication cookies are never included. The fixture is source-derived because runtime Swagger serialization was unavailable in the isolated finalization clone; Staff.Bff.Tests validates the paths, permissions, DTO shapes and response metadata against the controllers."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": [
    {
      "cookieAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/shipments/{id}/document-intakes": {
      "post": {
        "tags": [
          "Shipments"
        ],
        "operationId": "CreateDocumentIntake",
        "description": "Verifies the upload, creates or replays one attachment, consumes the upload session and submits one idempotent OCR job.",
        "x-required-permission": "documents:ingest",
        "parameters": [
          {
            "$ref": "#/components/parameters/ShipmentId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Staff_CreateDocumentIntakeBody"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Document intake accepted or replayed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Staff_DocumentIntakeHttpResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Problem400_6"
          },
          "401": {
            "$ref": "#/components/responses/Problem401_1"
          },
          "404": {
            "$ref": "#/components/responses/Problem404_7"
          },
          "409": {
            "$ref": "#/components/responses/Problem409_8"
          },
          "422": {
            "$ref": "#/components/responses/Problem422_4"
          },
          "500": {
            "$ref": "#/components/responses/Problem500_13"
          },
          "503": {
            "$ref": "#/components/responses/Problem503_9"
          }
        },
        "x-error-contract": {
          "400": [
            {
              "code": "INVALID_REQUEST",
              "retryable": false
            }
          ],
          "401": [
            {
              "code": "TENANT_CONTEXT_REQUIRED",
              "retryable": false
            }
          ],
          "404": [
            {
              "code": "UPLOAD_NOT_FOUND",
              "retryable": false
            },
            {
              "code": "UPLOAD_OBJECT_NOT_FOUND",
              "retryable": false
            },
            {
              "code": "UPLOAD_TENANT_MISMATCH",
              "retryable": false
            },
            {
              "code": "DOCUMENT_INTAKE_NOT_FOUND",
              "retryable": false
            },
            {
              "code": "DOCUMENT_NOT_FOUND",
              "retryable": false
            }
          ],
          "409": [
            {
              "code": "UPLOAD_EXPIRED",
              "retryable": false
            },
            {
              "code": "UPLOAD_IDEMPOTENCY_CONFLICT",
              "retryable": false
            },
            {
              "code": "UPLOAD_NOT_VERIFIED",
              "retryable": false
            },
            {
              "code": "IDEMPOTENCY_CONFLICT",
              "retryable": false
            },
            {
              "code": "INVALID_STATE_TRANSITION",
              "retryable": false
            },
            {
              "code": "UPLOAD_VERIFICATION_IN_PROGRESS",
              "retryable": true
            },
            {
              "code": "UPLOAD_NOT_READY",
              "retryable": false
            }
          ],
          "422": [
            {
              "code": "UPLOAD_INVALID",
              "retryable": false
            },
            {
              "code": "UPLOAD_MIME_MISMATCH",
              "retryable": false
            },
            {
              "code": "UPLOAD_SIZE_MISMATCH",
              "retryable": false
            },
            {
              "code": "UPLOAD_HASH_MISMATCH",
              "retryable": false
            },
            {
              "code": "UPLOAD_CONTENT_MISMATCH",
              "retryable": false
            },
            {
              "code": "UPLOAD_SIZE_EXCEEDED",
              "retryable": false
            }
          ],
          "500": [
            {
              "code": "DOCUMENT_CONTRACT_ERROR",
              "retryable": false
            }
          ],
          "503": [
            {
              "code": "DOCUMENT_OCR_UNAVAILABLE",
              "retryable": true
            },
            {
              "code": "SHIPMENT_WORKFLOW_UNAVAILABLE",
              "retryable": true
            }
          ]
        }
      }
    },
...
    "/api/v1/documents/shipment/{id}": {
      "get": {
        "tags": [
          "Documents"
        ],
        "operationId": "GetShipmentDocumentStatus",
        "x-required-permission": "documents:read",
        "parameters": [
          {
            "$ref": "#/components/parameters/DocumentId"
          }
        ],
        "responses": {
          "200": {
            "description": "OCR status and extraction summary",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Staff_UnifiedDocumentStatusResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Problem401_1"
          },
          "404": {
            "$ref": "#/components/responses/Problem404_11"
          },
          "503": {
            "$ref": "#/components/responses/Problem503_5"
          }
        },
        "x-error-contract": {
          "401": [
            {
              "code": "TENANT_CONTEXT_REQUIRED",
              "retryable": false
            }
          ],
          "404": [
            {
              "code": "DOCUMENT_NOT_FOUND",
              "retryable": false
            }
          ],
          "503": [
            {
              "code": "DOCUMENT_OCR_UNAVAILABLE",
              "retryable": true
            }
          ]
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": ".Aurora.Auth"
      }
    },
    "parameters": {
      "ShipmentId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
      "DocumentId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      }
    },
    "schemas": {
...
      "Staff_CreateDocumentIntakeBody": {
        "type": "object",
        "required": [
          "uploadId",
          "documentTypeHint",
          "idempotencyKey"
        ],
        "properties": {
          "uploadId": {
            "type": "string",
            "format": "uuid"
          },
          "documentTypeHint": {
            "type": "string"
          },
          "idempotencyKey": {
            "type": "string"
          }
        }
      },
      "Staff_DocumentIntakeHttpResponse": {
        "type": "object",
        "required": [
          "intakeId",
          "shipmentId",
          "documentId",
          "uploadId",
          "ocrJobId",
          "status",
          "intakeStatus",
          "fileName",
          "createdAt",
          "retryable"
        ],
        "properties": {
          "intakeId": {
            "type": "string",
            "format": "uuid"
          },
          "shipmentId": {
            "type": "string",
            "format": "uuid"
          },
          "documentId": {
            "type": "string",
            "format": "uuid"
          },
          "uploadId": {
            "type": "string",
            "format": "uuid"
          },
          "ocrJobId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string",
            "enum": [
              "RECEIVED",
              "PROCESSING",
              "READY",
              "NEEDS_REVIEW",
              "REJECTED",
              "FAILED",
              "CANCELLED"
            ]
          },
          "stage": {
            "type": "string",
            "nullable": true,
            "enum": [
              "QUEUED",
              "EXTRACTING",
              "HUMAN_REVIEW",
              "COMPLETED",
              "ERROR"
            ]
          },
          "intakeStatus": {
            "type": "string",
            "enum": [
              "PENDING_ATTACHMENT",
              "PENDING_OCR",
              "SUBMITTED",
              "FAILED_RETRYABLE"
            ]
          },
          "fileName": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "retryable": {
            "type": "boolean"
          },
          "retryAfterSeconds": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        }
      },
      "Staff_UnifiedDocumentStatusResponse": {
        "type": "object",
        "required": [
          "id",
          "documentType",
          "status",
          "fileName",
          "needsReview"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "documentType": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "RECEIVED",
              "PROCESSING",
              "READY",
              "NEEDS_REVIEW",
              "REJECTED",
              "FAILED",
              "CANCELLED"
            ]
          },
          "stage": {
            "type": "string",
            "nullable": true
          },
          "fileName": {
            "type": "string"
          },
          "needsReview": {
            "type": "boolean"
          },
          "confidence": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "normalizedJson": {
            "type": "string",
            "nullable": true
          },
          "errorCode": {
            "type": "string",
            "nullable": true
          },
          "errorMessage": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
...
    },
    "responses": {
      "Problem401_1": {
        "description": "The authenticated caller has no trusted tenant context.",
        "content": {
          "application/problem+json": {
            "schema": {
              "oneOf": [
                {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "type": "object",
                      "required": [
                        "status",
                        "code",
                        "retryable"
                      ],
                      "properties": {
                        "status": {
                          "type": "integer",
                          "format": "int32",
                          "enum": [
                            401
                          ]
                        },
                        "code": {
                          "type": "string",
                          "enum": [
                            "TENANT_CONTEXT_REQUIRED"
                          ]
                        },
                        "retryable": {
                          "type": "boolean",
                          "enum": [
                            false
                          ]
                        }
                      }
                    }
                  ]
                }
              ]
            },
            "examples": {
              "TENANT_CONTEXT_REQUIRED": {
                "summary": "TENANT_CONTEXT_REQUIRED",
                "value": {
                  "type": "https://httpstatuses.io/401",
                  "title": "TENANT_CONTEXT_REQUIRED",
                  "status": 401,
                  "detail": "A trusted tenant context is required for this resource.",
                  "code": "TENANT_CONTEXT_REQUIRED",
                  "retryable": false
                }
              }
            }
          }
        }
      },
...
      "Problem400_6": {
        "description": "Invalid request. The caller must correct the request before retrying.",
        "content": {
          "application/problem+json": {
            "schema": {
              "oneOf": [
                {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "type": "object",
                      "required": [
                        "status",
                        "code",
                        "retryable"
                      ],
                      "properties": {
                        "status": {
                          "type": "integer",
                          "format": "int32",
                          "enum": [
                            400
                          ]
                        },
                        "code": {
                          "type": "string",
                          "enum": [
                            "INVALID_REQUEST"
                          ]
                        },
                        "retryable": {
                          "type": "boolean",
                          "enum": [
                            false
                          ]
                        }
                      }
                    }
                  ]
                }
              ]
            },
            "examples": {
              "INVALID_REQUEST": {
                "summary": "INVALID_REQUEST",
                "value": {
                  "type": "https://httpstatuses.io/400",
                  "title": "INVALID_REQUEST",
                  "status": 400,
                  "detail": "Invalid request. The caller must correct the request before retrying.",
                  "code": "INVALID_REQUEST",
                  "retryable": false
                }
              }
            }
          }
        }
      },
...
    }
  }
}